requirements.txt
click==8.1.7
colorama==0.4.6
Flask==2.2.5
importlib-metadata==6.7.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
mysql-connector-python==8.0.33
protobuf==3.20.3
python-dotenv==0.21.1
typing-extensions==4.7.1
Werkzeug==2.2.3
zipp==3.15.0
